refactor(mxfp_moe): unify global scalar access into common
Consolidate the two duplicate global-scalar-access styles used by the
mxfp_moe stage kernels into one canonical set in mxfp4_gemm_common.py,
shared by both gemm1 and gemm2.

- Relocate stateless scalar-access helpers (_global_i32_ptr / _global_i32_at
  / _global_i32_load / _global_scalar_tiles / _scalar_store / _layout_idx)
  and stateless numeric helpers (_pkmax_u16 / _inline_e8m0 / _silu_mul_batch,
  plus LOG2E) from gemm1 into common; gemm1 imports them back. Pure
  relocation: gemm1 final ISA is byte-identical.
- Migrate gemm2's raw scalar reads (cumsum0 x2, eids) from
  llvm.load(T.i32, _global_ptr1(...)) to common's _global_i32_at(...).
  The stids/sweights invariant scalar loads are left raw (they carry the
  !invariant.load hint that _global_i32_at does not emit). gemm2 final ISA
  is byte-identical.
- Drop the now-dead _global_ptr1 helper (no remaining callers; _gep1 /
  _global_base_ptr1 stay for non-scalar pointer needs).

Verified codegen-equivalent by diffing dumped final ISA (21_final_isa.s)
for both kernels: identical vs 6ebcda9c. Cold correctness (cache off,
RUN_QUANT=1): test_mxfp_moe_variants 6/6; a4w4+a8w4 FP4-M/FP4-L e2e
46 passed / 14 skipped.
